[WinEH] Fix try scopes leaking to caller on inline
This fixes issue #164169

When inlining functions compiled with -EHa, try scope terminators might need to be inserted before inlined returns.
This prevents leaking try scopes over to the caller.
Try scopes can be ended before a ret due to the unwinder forwarding exceptions in the seh epilog to the caller.